Edit / Delete / Create frontend implementation for Sports

+++ backend/routes/sports.js
@@ -1,5 +1,7 @@
 import express from 'express';
-import { getAllSports, getSpecificSport, deleteSport} from '../services/sportService.js';
+import { authenticateToken } from '../middleware/auth.js';
+import { deleteContentLimiter } from '../middleware/rate_limiter.js';
+import { getAllSports, getSpecificSport, deleteSport, createSport, editSport } from '../services/sportService.js';
 
 const router = express.Router();
 
@@ -34,14 +36,69 @@ router.get('/:sportName', async(req, res) => {
     }
 })
 
-router.delete('/:sportName', async(req, res) => {
+// Delete a Sport (ADMINS Only)
+router.delete('/:sportName', authenticateToken, async(req, res) => {
     try {
-        const {sportName} = req.params;
-        await deleteSport(req.pgClient, sportName);
+        const { sportName } = req.params;
+
+        if (!req.user) {
+            return res.status(401).json({ error: 'Authentication required' });
+        }
+        
+        const { username } = req.user;
+        console.log("In backend: ", username);
+        
+        if (username !== process.env.ADMIN_USERNAME) {
+            return res.status(403).json({ error: 'Unauthorized. Only ADMINs can delete sports.' });
+        }
+
+        const result = await deleteSport(req.pgClient, sportName);
         
         res.status(201).json({ message: 'Sport successfully deleted'})
     } catch (error) {
         console.error('Error deleting sport:', error);
+        res.status(500).json({ error: error.message });
+    }
+})
+
+router.post('/', authenticateToken, async (req, res) => {
+    try {
+        const sportData = req.body;
+
+        console.log(sportData);
+
+        if (!req.user) {
+            return res.status(401).json({ error: 'Authentication required' });
+        }
+        
+        const { username } = req.user;
+        console.log("In backend: ", username);
+
+        const result = await createSport(req.pgClient, sportData);
+        res.status(201).json({ message: 'Sport Successfully Created' });
+    } catch (error) {
+        console.error('Error creating sport:', error);
+        res.status(500).json({ error: error.message });
+    }
+})
+
+router.put('/:sportName', authenticateToken, async (req, res) => {
+    try {
+        const sportData = req.body;
+
+        console.log(sportData);
+
+        if (!req.user) {
+            return res.status(401).json({ error: 'Authentication required' });
+        }
+        
+        const { username } = req.user;
+        console.log("In backend: ", username);
+
+        const result = await editSport(req.pgClient, sportData);
+        res.status(201).json({ message: 'Sport Successfully Updated' });
+    } catch (error) {
+        console.error('Error editing sport:', error);
         res.status(500).json({ error: error.message });
     }
 })
